refactor : remove unnecessary types and add ajust in some parts of code
