+++ src/presentation/controllers/manufacturer.ts
@@ -2,14 +2,14 @@ import { Request } from "express";
 import { IAddManufacturerUseCase } from "../../application/usecases/manufacturer";
 import { Controller, HttpResponse } from "../../domain/controller";
 import { HttpStatusCode } from "../../domain/http";
-import { CreateManufacturerInput } from "../../domain/manufacturer";
+import { Manufacturer } from "../../domain/manufacturer";
 
 export class AddManufacturerController implements Controller {
   constructor(readonly addManufacturerUseCase: IAddManufacturerUseCase) {}
 
   async handle(request: Request): Promise<HttpResponse> {
     try {
-      const data = request.body as CreateManufacturerInput;
+      const data = request.body as Manufacturer;
       const manufacturer = await this.addManufacturerUseCase.execute(data);
       return {
         msg: "Manufacturer created successfully",
